+++ Cargo.toml
@@ -6,3 +6,4 @@ edition = "2018"
 
 [dependencies]
 scrap = { git = "https://github.com/ncatelli/scrap", tag = "v0.1.1" }
+parcel = { git = "https://github.com/ncatelli/parcel", tag = "v1.9.0" }
