+++ src/cpu/mos6502/mod.rs
@@ -1,0 +1,173 @@
+use crate::{
+    address_map::{
+        memory::{Memory, ReadWrite},
+        AddressMap, Addressable,
+    },
+    cpu::{
+        register::{GeneralPurpose, ProcessorStatus, ProgramCounter, Register, StackPointer},
+        CPU,
+    },
+};
+
+/// MOS6502 represents the 6502 CPU
+#[derive(Debug)]
+pub struct MOS6502 {
+    address_map: AddressMap<u16>,
+    pub acc: GeneralPurpose,
+    pub x: GeneralPurpose,
+    pub y: GeneralPurpose,
+    pub sp: StackPointer,
+    pub pc: ProgramCounter,
+    pub ps: ProcessorStatus,
+}
+
+impl MOS6502 {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// instantiates a new MOS6502 with a provided address_map.
+    pub fn with_addressmap(am: AddressMap<u16>) -> Self {
+        let mut cpu = Self::default();
+        cpu.address_map = am;
+        cpu
+    }
+
+    /// emulates the reset process of the CPU.
+    pub fn reset(self) -> Self {
+        let mut cpu = MOS6502::with_addressmap(self.address_map);
+        let lsb: u8 = cpu.address_map.read(0x7ffc);
+        let msb: u8 = cpu.address_map.read(0x7ffd);
+
+        cpu.pc = ProgramCounter::default().write(u16::from_le_bytes([lsb, msb]));
+        cpu
+    }
+}
+
+impl Default for MOS6502 {
+    fn default() -> Self {
+        Self {
+            address_map: AddressMap::new()
+                .register(
+                    0x0000..0x00FF,
+                    Box::new(Memory::<ReadWrite>::new(0x0000, 0x00FF)),
+                )
+                .unwrap()
+                .register(
+                    0x0100..0x01FF,
+                    Box::new(Memory::<ReadWrite>::new(0x0100, 0x01FF)),
+                )
+                .unwrap(),
+            acc: GeneralPurpose::default(),
+            x: GeneralPurpose::default(),
+            y: GeneralPurpose::default(),
+            sp: StackPointer::default(),
+            pc: ProgramCounter::default(),
+            ps: ProcessorStatus::default(),
+        }
+    }
+}
+
+impl<MOS6502> CPU<MOS6502> for MOS6502 {
+    fn step(cpu: MOS6502) -> MOS6502 {
+        cpu
+    }
+}
+
+#[derive(Clone, Copy, PartialEq, Debug)]
+pub enum Mnemonic {
+    // Load-Store
+    LDA,
+    LDX,
+    LDY,
+    STA,
+    STX,
+    STY,
+
+    // Arithmetic
+    ADC,
+    SBC,
+    INC,
+    INX,
+    INY,
+    DEC,
+    DEX,
+    DEY,
+
+    // Shift and Rotate
+    ASL,
+    LSR,
+    ROL,
+    ROR,
+    AND,
+    ORA,
+    EOR,
+
+    // Compare and Test Bit
+    CMP,
+    CPX,
+    CPY,
+    BIT,
+
+    // Branch
+    BCC,
+    BCS,
+    BNE,
+    BEQ,
+    BPL,
+    BMI,
+    BVC,
+    BVS,
+
+    // Transfer
+    TAX,
+    TXA,
+    TAY,
+    TYA,
+    TSX,
+    TXS,
+
+    // Stack
+    PHA,
+    PLA,
+    PHP,
+    PLP,
+
+    // Subroutines and Jump
+    JMP,
+    JSR,
+    RTS,
+    RTI,
+
+    // Set and Clear
+    CLC,
+    SEC,
+    CLD,
+    SED,
+    CLI,
+    SEI,
+    CLV,
+
+    // Misc
+    BRK,
+    NOP,
+}
+
+/// AddressMode captures the Address mode type with a corresponding
+/// operand of the appropriate bit length.
+#[derive(Clone, Copy, PartialEq, Debug)]
+pub enum AddressMode {
+    Accumulator,
+    Implied,
+    Immediate(u8),
+    Absolute(u16),
+    ZeroPage(u8),
+    Relative(i8),
+    Indirect(u16),
+    AbsoluteIndexedWithX(u16),
+    AbsoluteIndexedWithY(u16),
+    ZeroPageIndexedWithX(u8),
+    ZeroPageIndexedWithY(u8),
+    IndexedIndirect(u8),
+    IndirectIndexed(u8),
+}
